refactor the factory class
